+++ src/player.cpp
@@ -2527,7 +2527,7 @@ void YoshiEat(const int A)
             {
                 if(n.Type == 91)
                 {
-                    if(!NPCNoYoshi[n.Special])
+                    if(!NPCNoYoshi[(int)n.Special])
                     {
                         PlaySound(SFX_Grab);
                         n.Generator = false;
